+++ packages/styles/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @foundation-ui/styles
 
+## 1.4.1
+
+### Patch Changes
+
+- Export component hooks
+
 ## 1.4.0
 
 ### Minor Changes